Moved classes to appropriate packages
