+++ src/main/java/com/michael/container/health/listener/HealthQueueListener.java
@@ -1,4 +1,4 @@
-package com.michael.container.registry.cache.listener.queue;
+package com.michael.container.health.listener;
 
 import com.michael.container.health.service.HealthCheckService;
 import com.michael.container.registry.cache.listener.key.KeyListener;
