+++ src/main/java/com/michael/container/notifications/fsm/status/BeginState.java
@@ -1,9 +1,9 @@
-package com.michael.container.registry.cache.listener.status;
+package com.michael.container.notifications.fsm.status;
 
 import com.michael.container.notifications.enums.NotificationType;
 import com.michael.container.notifications.model.ServiceNotificationRequest;
 import com.michael.container.notifications.service.NotificationService;
-import com.michael.container.registry.enums.StatusStateEvent;
+import com.michael.container.notifications.enums.StatusStateEvent;
 import com.michael.container.registry.model.StatusChangeEvent;
 import jakarta.annotation.Nonnull;
 import org.springframework.beans.factory.annotation.Qualifier;
