+++ src/main/java/com/michael/container/notifications/fsm/status/MaintenanceState.java
@@ -1,19 +1,19 @@
-package com.michael.container.registry.cache.listener.status;
+package com.michael.container.notifications.fsm.status;
 
 import com.michael.container.notifications.enums.NotificationType;
 import com.michael.container.notifications.model.ServiceNotificationRequest;
 import com.michael.container.notifications.service.NotificationService;
-import com.michael.container.registry.enums.StatusStateEvent;
+import com.michael.container.notifications.enums.StatusStateEvent;
 import com.michael.container.registry.model.StatusChangeEvent;
 import jakarta.annotation.Nonnull;
 import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.stereotype.Component;
 
 @Component
-public class FixState implements StatusChange {
+public class MaintenanceState implements StatusChange {
   private final NotificationService notificationService;
 
-  public FixState(
+  public MaintenanceState(
       @Qualifier("generalNotificationService") NotificationService notificationService) {
     this.notificationService = notificationService;
   }
@@ -31,6 +31,6 @@ public class FixState implements StatusChange {
 
   @Override
   public StatusStateEvent getStatusStateEvent() {
-    return StatusStateEvent.FIX;
+    return StatusStateEvent.MAINTENANCE;
   }
 }
