+++ src/main/java/com/michael/container/notifications/fsm/status/StatusChangeOrchestrator.java
@@ -1,7 +1,7 @@
-package com.michael.container.registry.cache.listener.status;
+package com.michael.container.notifications.fsm.status;
 
 import com.michael.container.registry.enums.Status;
-import com.michael.container.registry.enums.StatusStateEvent;
+import com.michael.container.notifications.enums.StatusStateEvent;
 import com.michael.container.registry.model.StatusChangeEvent;
 import jakarta.annotation.Nonnull;
 import java.util.List;
