+++ src/main/java/com/michael/container/utils/ContainerConstants.java
@@ -4,4 +4,5 @@ public class ContainerConstants {
   private ContainerConstants() {}
 
   public static final Long INSTANCE_ENTITY_MAINTENANCE_TIME_TO_LIVE = 5400L;
+  public static final Long INSTANCE_ENTITY_DEFAULT_TIME_TO_LIVE = 20L;
 }
